exercise 4.a testing node application

+++ controllers/blogs.js
@@ -8,8 +8,7 @@ require("express-async-errors")
 blogRouter.get('/', async (request, response) => {
     const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
     response.json(blogs.map(blog => blog.toJSON()))
-}
-)
+})
 // blogRouter.post('/', async (request, response, next) => {
 //     const blog = request.body
 //     const newBlog = new Blog({
@@ -25,8 +24,8 @@ blogRouter.get('/', async (request, response) => {
 //     }
 // })
 
-blogRouter.get('/:id', async (request, response) => {
-    const blog = await Blog.findById(request.params.id)
+blogRouter.get('/:id', async (request, response, next) => {
+    const blog = await Blog.findById(request.params.id).catch(error => next(error))
     if (blog) {
         response.status(200).json(blog.toJSON())
     }
@@ -35,24 +34,14 @@ blogRouter.get('/:id', async (request, response) => {
     }
 })
 
-blogRouter.delete('/:id', async (request, response) => {
-    const token = request.token
-    const decodedToken = jwt.verify(token, process.env.SECRET)
-    if (!token || !decodedToken.id) {
-        return response.status(401).json({ error: 'token missing or invalid' })
-    }
-    const deletedBlog = await Blog.findOneAndDelete({ _id: request.params.id })
-    response.status(204).json(deletedBlog)
-})
-
-blogRouter.post('/', async (request, response) => {
+blogRouter.post('/', async (request, response, next) => {
     const body = request.body
     const token = request.token
     const decodedToken = jwt.verify(token, process.env.SECRET)
     if (!token || !decodedToken.id) {
         return response.status(401).json({ error: 'token missing or invalid' })
     }
-    const user = await User.findById(decodedToken.id)
+    const user = await User.findById(decodedToken.id).catch(error => next(error))
     const blog = new Blog({
         title: body.title,
         author: body.author,
@@ -61,25 +50,40 @@ blogRouter.post('/', async (request, response) => {
         user: user._id
     })
 
-    const savedBlog = await blog.save()
+    const savedBlog = await blog.save().catch(error => next(error))
     user.blogs = user.blogs.concat(savedBlog._id)
-    await user.save()
+    await user.save().catch(error => next(error))
     response.status(201).json(savedBlog.toJSON())
 })
 
-blogRouter.put('/:id', async (request, response) => {
+blogRouter.delete('/:id', async (request, response, next) => {
+    const token = request.token
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+    if (!token || !decodedToken.id) {
+        return response.status(401).json({ error: 'token missing or invalid' })
+    }
+    const deletedBlog = await Blog.findOneAndDelete({ _id: request.params.id }).catch(error => next(error))
+    response.status(204).json(deletedBlog)
+})
+
+blogRouter.put('/:id', async (request, response, next) => {
     const body = request.body
     const token = request.token
     const decodedToken = jwt.verify(token, process.env.SECRET)
     if (!token || !decodedToken.id) {
         return response.status(401).json({ error: 'token missing or invalid' })
     }
-    const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, body, { new: true })
+    const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, body, { new: true }).catch(error => next(error))
     response.status(201).json(updatedBlog.toJSON())
 })
 
 
 blogRouter.post('/reset', async (request, response) => {
+    const token = request.token
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+    if (!token || !decodedToken.id) {
+        return response.status(401).json({ error: 'token missing or invalid' })
+    }
     await Blog.deleteMany({})
     await User.deleteMany({})
